requirements.txt
fire
langchain-core
langchain-openai
openai
python-dotenv
python-dotenv-vault
logging